🍕 Decomposed Index to modules, updated namings in router
