+++ src/frontend/src/layouts/HeaderLayout.vue
@@ -1,0 +1,28 @@
+<template>
+  <header class="header">
+    <div class="header__logo">
+      <a href="/" class="logo">
+        <img
+          src="@/assets/img/logo.svg"
+          alt="V!U!E! Pizza logo"
+          width="90"
+          height="40"
+        />
+      </a>
+    </div>
+    <div class="header__cart">
+      <a href="cart.html">0 ₽</a>
+    </div>
+    <div class="header__user">
+      <a href="#" class="header__login"><span>Войти</span></a>
+    </div>
+  </header>
+</template>
+
+<script>
+export default {
+  name: "HeaderLayout",
+};
+</script>
+
+<style></style>
